Moved linking and social login to different files. Also added tags to user.yaml
